Move ConsistencyCheckJobType to scenario module
