+++ kernel/data-pipeline/scenario/consistencycheck/src/main/resources/META-INF/services/org.apache.shardingsphere.data.pipeline.spi.job.JobType
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.data.pipeline.core.job.type.ConsistencyCheckJobType
+org.apache.shardingsphere.data.pipeline.scenario.consistencycheck.ConsistencyCheckJobType
